+++ include/triton-shared/Conversion/CMakeLists.txt
@@ -2,5 +2,5 @@ add_subdirectory(TritonToLinalg)
 add_subdirectory(TritonToLinalgExperimental)
 add_subdirectory(TritonToStructured)
 add_subdirectory(TritonArithToLinalg)
-add_subdirectory(FoldUnstructuredPtr)
+add_subdirectory(TritonToUnstructured)
 add_subdirectory(StructuredToMemref)
