+++ include/triton-shared/Conversion/TritonToUnstructured/CMakeLists.txt
@@ -1,0 +1,3 @@
+set(LLVM_TARGET_DEFINITIONS Passes.td)
+mlir_tablegen(Passes.h.inc -gen-pass-decls --name TritonToUnstructured)
+add_public_tablegen_target(TritonToUnstructuredConversionPassIncGen)
