+++ include/triton-shared/Conversion/TritonToUnstructured/Passes.td
@@ -1,11 +1,11 @@
-#ifndef FOLD_UNSTRUCTURED_PTR_CONVERSION_PASSES
-#define FOLD_UNSTRUCTURED_PTR_CONVERSION_PASSES
+#ifndef TRITON_TO_UNSTRUCTURED_CONVERSION_PASSES
+#define TRITON_TO_UNSTRUCTURED_CONVERSION_PASSES
 
 include "mlir/Pass/PassBase.td"
 
-def FoldUnstructuredPtr : Pass<"fold-unstructured-ptr", "mlir::ModuleOp"> {
+def TritonToUnstructured : Pass<"triton-to-unstructured", "mlir::ModuleOp"> {
   let summary = "Transforms tt.addptr ops into offset accumulation ops";
-  let constructor = "triton::createFoldUnstructuredPtrPass()";
+  let constructor = "triton::createTritonToUnstructuredPass()";
   let options = [
       Option<"offsetBitWidth", "offset-bit-width", "size_t", /*default*/"32",
              "Bitwidth used for the starting offset of each pointer">
