+++ lib/Conversion/CMakeLists.txt
@@ -1,6 +1,6 @@
-add_subdirectory(FoldUnstructuredPtr)
 add_subdirectory(TritonToLinalg)
 add_subdirectory(TritonToLinalgExperimental)
 add_subdirectory(TritonToStructured)
+add_subdirectory(TritonToUnstructured)
 add_subdirectory(TritonArithToLinalg)
 add_subdirectory(StructuredToMemref)
