+++ test/Conversion/TritonToUnstructured/make_tensor_ptr_structured.mlir
@@ -1,4 +1,4 @@
-// RUN: triton-shared-opt --fold-unstructured-ptr %s | FileCheck %s
+// RUN: triton-shared-opt --triton-to-structured --triton-to-unstructured %s | FileCheck %s
 
 module {
   tt.func public @add_ptr_into_make_block_ptr(%arg0: !tt.ptr<bf16>, %arg1: !tt.ptr<bf16>) attributes {noinline = false} {
@@ -21,20 +21,18 @@ module {
 }
 
 // CHECK:         tt.func public @add_ptr_into_make_block_ptr([[PARAM_0_:%.+]]: !tt.ptr<bf16>, [[PARAM_1_:%.+]]: !tt.ptr<bf16>) attributes {noinline = false} {
-// CHECK-DAG:       [[CST_0_:%.+]] = arith.constant 0 : i32
+// CHECK-DAG:       [[CST_0_:%.+]] = arith.constant 0 : index
+// CHECK-DAG:       [[CST_1_:%.+]] = arith.constant 1 : index
+// CHECK-DAG:       [[CST_512_:%.+]] = arith.constant 512 : index
+// CHECK-DAG:       [[CST_64_:%.+]] = arith.constant 64 : index
 // CHECK-DAG:       [[CST_32768_:%.+]] = arith.constant 32768 : i64
-// CHECK-DAG:       [[CST_1_:%.+]] = arith.constant 1 : i64
-// CHECK-DAG:       [[CST_64_:%.+]] = arith.constant 64 : i64
-// CHECK-DAG:       [[CST_512_:%.+]] = arith.constant 512 : i64
 // CHECK-DAG:       [[VAR_0_:%.+]] = tt.get_program_id x : i32
 // CHECK:           [[VAR_1_:%.+]] = arith.extsi [[VAR_0_]] : i32 to i64
 // CHECK:           [[VAR_2_:%.+]] = arith.muli [[VAR_1_]], [[CST_32768_]] : i64
-// CHECK:           [[VAR_3_:%.+]] = "tts.make_unstructured_tptr"([[PARAM_0_]], [[VAR_2_]]) : (!tt.ptr<bf16>, i64) -> !tt.ptr<bf16>
-// CHECK-DAG:       [[VAR_4_:%.+]] = tt.make_tensor_ptr [[VAR_3_]], {{.}}[[CST_512_]], [[CST_64_]]{{.}}, {{.}}[[CST_64_]], [[CST_1_]]{{.}}, {{.}}[[CST_0_]], [[CST_0_]]{{.}} {order = array<i32: 1, 0>} : <tensor<512x64xbf16>>
-// CHECK-DAG:       [[VAR_5_:%.+]] = "tts.make_unstructured_tptr"([[PARAM_1_]], [[VAR_2_]]) : (!tt.ptr<bf16>, i64) -> !tt.ptr<bf16>
-// CHECK-NOT: separator of consecutive DAGs
-// CHECK-DAG:       [[VAR_6_:%.+]] = tt.make_tensor_ptr [[VAR_5_]], {{.}}[[CST_512_]], [[CST_64_]]{{.}}, {{.}}[[CST_64_]], [[CST_1_]]{{.}}, {{.}}[[CST_0_]], [[CST_0_]]{{.}} {order = array<i32: 1, 0>} : <tensor<512x64xbf16>>
-// CHECK-DAG:       [[LOAD_VAR_4_MEM_:%.+]] = tt.load [[VAR_4_]] : !tt.ptr<tensor<512x64xbf16>>
-// CHECK:           tt.store [[VAR_6_]], [[LOAD_VAR_4_MEM_]] : !tt.ptr<tensor<512x64xbf16>>
+// CHECK:           [[VAR_3_:%.+]] = arith.index_cast [[VAR_2_]] : i64 to index
+// CHECK-DAG:       [[VAR_4_:%.+]] = tts.make_tptr [[PARAM_0_]] to sizes: [512, 64], strides: {{.}}[[CST_64_]], [[CST_1_]]{{.}}, offsets: {{.}}[[VAR_3_]], [[CST_0_]]{{.}}, shape: {{.}}[[CST_512_]], [[CST_64_]]{{.}}, order: [1, 0] : <bf16> to !tt.ptr<tensor<512x64xbf16>>
+// CHECK-DAG:       [[VAR_5_:%.+]] = tts.make_tptr [[PARAM_1_]] to sizes: [512, 64], strides: {{.}}[[CST_64_]], [[CST_1_]]{{.}}, offsets: {{.}}[[VAR_3_]], [[CST_0_]]{{.}}, shape: {{.}}[[CST_512_]], [[CST_64_]]{{.}}, order: [1, 0] : <bf16> to !tt.ptr<tensor<512x64xbf16>>
+// CHECK:           [[VAR_6_:%.+]] = "tts.load"([[VAR_4_]]) <{operandSegmentSizes = array<i32: 1, 0, 0>, static_mask_dims = array<i64>}> : (!tt.ptr<tensor<512x64xbf16>>) -> tensor<512x64xbf16>
+// CHECK:           "tts.store"([[VAR_5_]], [[VAR_6_]]) <{static_mask_dims = array<i64>}> : (!tt.ptr<tensor<512x64xbf16>>, tensor<512x64xbf16>) -> ()
 // CHECK:           tt.return
 // CHECK:         }
